RPI/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.1
fonttools==4.47.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
mariadb==1.1.9
MarkupSafe==2.1.3
matplotlib==3.8.2
multidict==6.0.4
ntplib==0.4.0
numpy==1.26.3
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
python-telegram-bot==20.7
six==1.16.0
sniffio==1.3.0
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.4
